requirements.txt
Django==1.5
Fabric==1.6
distribute==0.6.30
MySQL-python==1.2.3
South==0.7.6
argparse==1.2.1
dj-database-url==0.2.1
pycrypto==2.6
ssh==1.8.0
wsgiref==0.1.2
django-bootstrap-toolkit
python-daemon